New Audio Peer, Reactive object and Efficient Movement Sxripts
The audio peer has be re written to better suit the experience.
I have made a object  that reacts to the music, which may or may not be temporary.
I have written a new script that allows for multiple behaviours from one object and is more efficient overall.
I imported the song that Imran has made for the experience and deleted the copyrighted music.